HUGE Renderer refactor - Added RenderCommandBuffer class - SceneRenderer is now intended to be instantiated - All (wip) Renderer commands are now written to a RenderCommandBuffer object - Added Asset Viewer panel (major WIP) - Added second viewport (Edit->Second Viewport) - OpenGL is currently broken (and will probably be removed)

+++ Engine/Source/Renderer/Backend/Vulkan/VulkanShader.hpp
@@ -60,7 +60,7 @@ namespace Vanta {
 			if (m_ShaderDescriptorSets.size() < set)
 				return 0;
 
-			return m_ShaderDescriptorSets[set].UniformBuffers.size();
+			return (uint32_t)m_ShaderDescriptorSets[set].UniformBuffers.size();
 		}
 
 		struct ShaderDescriptorSet
