+++ Engine/Source/Renderer/Image.hpp
@@ -120,7 +120,7 @@ namespace Vanta {
 
 		inline uint32_t CalculateMipCount(uint32_t width, uint32_t height)
 		{
-			return std::floor(std::log2(glm::min(width, height))) + 1;
+			return (uint32_t)std::floor(std::log2(glm::min(width, height))) + 1;
 		}
 
 		inline uint32_t GetImageMemorySize(ImageFormat format, uint32_t width, uint32_t height)
