+++ Engine/Source/Renderer/VertexBuffer.hpp
@@ -81,7 +81,7 @@ namespace Vanta {
 
 		uint32_t GetStride() const { return m_Stride; }
 		const std::vector<VertexBufferElement>& GetElements() const { return m_Elements; }
-		uint32_t GetElementCount() const { return m_Elements.size(); }
+		uint32_t GetElementCount() const { return (uint32_t)m_Elements.size(); }
 
 		std::vector<VertexBufferElement>::iterator begin() { return m_Elements.begin(); }
 		std::vector<VertexBufferElement>::iterator end() { return m_Elements.end(); }
